requirements.txt
Flask
Flask-SQLAlchemy
pandas
google-cloud-storage
google-cloud-bigquery
pytz
fastapi
uvicorn
pyarrow